+++ oauth2_provider/contrib/ninja/__init__.py
@@ -1,4 +1,5 @@
 from .security import HttpOAuth2
+from .throttling import OAuth2ClientRateThrottle, OAuth2UserOrClientRateThrottle
 
 
-__all__ = ["HttpOAuth2"]
+__all__ = ["HttpOAuth2", "OAuth2ClientRateThrottle", "OAuth2UserOrClientRateThrottle"]