+++ oauth2_provider/contrib/rest_framework/__init__.py
@@ -7,3 +7,4 @@ from .permissions import (
     TokenHasScope,
     TokenMatchesOASRequirements,
 )
+from .throttling import OAuth2ClientRateThrottle, OAuth2UserOrClientRateThrottle